requirements.txt
appdirs==1.4.4
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
baron==0.10.1
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==4.0.3
django-cache-url==3.3.0
django-cors-headers==3.11.0
django-environs==0.1.1
djangorestframework==3.13.1
environs==9.5.0
marshmallow==3.14.1
<<<<<<< HEAD
mysqlclient==1.3.12
pycodestyle==2.8.0
=======
mysqlclient==2.1.0
pycodestyle==2.8.0
PyMySQL==1.0.2
>>>>>>> staging
python-dotenv==0.19.2
pytz==2021.3
redbaron==0.9.2
rply==0.7.8
sqlparse==0.4.2
toml==0.10.2
